requirements.txt
asgiref==3.9.1
Django==5.2.4
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.4.2
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
